Update README.md
This is a Sportsperson Face Recognition System which is trained with images of some famous players (Messi, Ronaldo, Neymar, Mbappe, MS Dhoni, Virat Kohli, Rohit Sharma). The user may drop the image down or select from files for recognization and the system will display the player with highest probability score as result along with the table of probability scores for other players as well.